:recycle: 코멘트 도메인 분리 (#263)
