--- application/src/docs/asciidoc/community/comments.adoc
@@ -1,51 +1,0 @@
-=== /community-comments
-
-==== 코멘트 조회
-
-.Request
-include::{snippets}/get-community-comments/http-request.adoc[]
-.Query Parameters
-include::{snippets}/get-community-comments/query-parameters.adoc[]
-.Response
-include::{snippets}/get-community-comments/http-response.adoc[]
-.Response Fields
-include::{snippets}/get-community-comments/response-fields.adoc[]
-
-==== 코멘트 작성
-
-.Request
-include::{snippets}/create-community-comment/http-request.adoc[]
-.Request Headers
-include::{snippets}/create-community-comment/request-headers.adoc[]
-.Request Fields
-include::{snippets}/create-community-comment/request-fields.adoc[]
-.Response
-include::{snippets}/create-community-comment/http-response.adoc[]
-.Response Fields
-include::{snippets}/create-community-comment/response-fields.adoc[]
-
-==== 코멘트 수정
-
-.Request
-include::{snippets}/update-community-comment/http-request.adoc[]
-.Request Headers
-include::{snippets}/update-community-comment/request-headers.adoc[]
-.Path Parameters
-include::{snippets}/update-community-comment/path-parameters.adoc[]
-.Response
-include::{snippets}/update-community-comment/http-response.adoc[]
-.Response Fields
-include::{snippets}/update-community-comment/response-fields.adoc[]
-
-==== 코멘트 삭제
-
-.Request
-include::{snippets}/delete-community-comment/http-request.adoc[]
-.Request Headers
-include::{snippets}/delete-community-comment/request-headers.adoc[]
-.Path Parameters
-include::{snippets}/delete-community-comment/path-parameters.adoc[]
-.Response
-include::{snippets}/delete-community-comment/http-response.adoc[]
-.Response Fields
-include::{snippets}/delete-community-comment/response-fields.adoc[]
