+++ application/src/main/kotlin/backend/team/ahachul_backend/api/comment/adapter/web/in/dto/DeleteCommentDto.kt
@@ -1,0 +1,8 @@
+package backend.team.ahachul_backend.api.comment.adapter.web.`in`.dto
+
+class DeleteCommentDto {
+
+    data class Response(
+        val id: Long,
+    )
+}
