--- application/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/comment/DeleteCommunityCommentDto.kt
@@ -1,8 +1,0 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment
-
-class DeleteCommunityCommentDto {
-
-    data class Response(
-        val id: Long,
-    )
-}
