--- application/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/post/DeleteCommunityPostDto.kt
@@ -1,14 +1,0 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post
-
-class DeleteCommunityPostDto {
-
-    data class Response(
-        val id: Long,
-    ) {
-        fun toCommand(): DeleteCommunityPostCommand {
-            return DeleteCommunityPostCommand(
-                id = id
-            )
-        }
-    }
-}
