+++ core/src/main/kotlin/backend/team/ahachul_backend/api/comment/application/command/CreateCommentCommand.kt
@@ -1,0 +1,8 @@
+package backend.team.ahachul_backend.api.comment.application.command
+
+class CreateCommentCommand(
+    val postId: Long,
+    val upperCommentId: Long?,
+    val content: String,
+) {
+}
