+++ core/src/main/kotlin/backend/team/ahachul_backend/api/comment/application/command/DeleteCommentCommand.kt
@@ -1,0 +1,6 @@
+package backend.team.ahachul_backend.api.comment.application.command
+
+class DeleteCommentCommand(
+    val id: Long,
+) {
+}
