+++ core/src/main/kotlin/backend/team/ahachul_backend/api/comment/application/command/GetCommentsCommand.kt
@@ -1,0 +1,6 @@
+package backend.team.ahachul_backend.api.comment.application.command
+
+class GetCommentsCommand(
+    val postId: Long
+) {
+}
