+++ core/src/main/kotlin/backend/team/ahachul_backend/api/comment/application/command/UpdateCommentCommand.kt
@@ -1,0 +1,7 @@
+package backend.team.ahachul_backend.api.comment.application.command
+
+class UpdateCommentCommand(
+    val id: Long,
+    val content: String,
+) {
+}
