+++ core/src/main/kotlin/backend/team/ahachul_backend/api/comment/domain/model/CommentType.kt
@@ -1,0 +1,5 @@
+package backend.team.ahachul_backend.api.comment.domain.model
+
+enum class CommentType {
+    CREATED, DELETED
+}
