+++ core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/CreateCommunityPostCommand.kt
