+++ core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/SearchCommunityPostCommand.kt
