--- core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/comment/CreateCommunityCommentCommand.kt
@@ -1,8 +1,0 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment
-
-class CreateCommunityCommentCommand(
-    val postId: Long,
-    val upperCommentId: Long?,
-    val content: String,
-) {
-}
