--- core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/comment/DeleteCommunityCommentCommand.kt
@@ -1,6 +1,0 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment
-
-class DeleteCommunityCommentCommand(
-    val id: Long,
-) {
-}
