--- core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/comment/UpdateCommunityCommentCommand.kt
@@ -1,7 +1,0 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment
-
-class UpdateCommunityCommentCommand(
-    val id: Long,
-    val content: String,
-) {
-}
