--- core/src/main/kotlin/backend/team/ahachul_backend/api/community/domain/model/CommunityCommentType.kt
@@ -1,5 +1,0 @@
-package backend.team.ahachul_backend.api.community.domain.model
-
-enum class CommunityCommentType {
-    CREATED, DELETED
-}
